@@ -4,13 +4,15 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/cdesiniotis/chord"
+	"time"
+
 	"github.com/cdesiniotis/chord/chordpb"
+	"github.com/cdesiniotis/chord"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"google.golang.org/grpc"
-	"time"
+	"google.golang.org/grpc/credentials/insecure"
 )
 
 func GetChordClient(addr string) (chordpb.ChordClient, error) {
@@ -19,7 +21,10 @@ func GetChordClient(addr string) (chordpb.ChordClient, error) {
 	defer cancel()
 
 	dialOpts := make([]grpc.DialOption, 0, 5)
-	dialOpts = append(dialOpts, grpc.WithInsecure(), grpc.WithBlock(), grpc.FailOnNonTempDialError(true))
+	dialOpts = append(dialOpts,
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+		grpc.WithBlock(),
+	)
 	conn, err := grpc.DialContext(ctx, addr, dialOpts...)
 	//conn, err := grpc.Dial(target, n.grpcOpts.dialOpts...)
 	if err != nil {
@@ -87,7 +92,7 @@ func readConfig(filename string, defaults map[string]interface{}) (*viper.Viper,
 
 func defaults() map[string]interface{} {
 	return map[string]interface{}{
-		"addr":	"0.0.0.0:8001",
+		"addr": "0.0.0.0:8001",
 	}
 }
 
