@@ -2,6 +2,7 @@ package chord
 
 import (
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials/insecure"
 )
 
 type Config struct {
@@ -19,13 +20,18 @@ type Config struct {
 
 	SuccessorListSize int
 
-	Logging 	bool
+	Logging          bool
+	EnableMetrics    bool
+	MetricsOutputDir string
 }
 
 func DefaultConfig(addr string, port int) *Config {
 	serverOpts := make([]grpc.ServerOption, 0, 5)
 	dialOpts := make([]grpc.DialOption, 0, 5)
-	dialOpts = append(dialOpts, grpc.WithInsecure(), grpc.WithBlock(), grpc.FailOnNonTempDialError(true)) //grpc.WithTimeout(5*time.Second)
+	dialOpts = append(dialOpts,
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+		grpc.WithBlock(),
+	)
 	return &Config{
 		KeySize:                  8,
 		Addr:                     addr,
@@ -37,14 +43,19 @@ func DefaultConfig(addr string, port int) *Config {
 		FixFingerInterval:        50,
 		CheckPredecessorInterval: 150,
 		SuccessorListSize:        2,
-		Logging:				  true,
+		Logging:                  true,
+		EnableMetrics:            false,
+		MetricsOutputDir:         "metrics",
 	}
 }
 
 func SetDefaultGrpcOpts(cfg *Config) *Config {
 	serverOpts := make([]grpc.ServerOption, 0, 5)
 	dialOpts := make([]grpc.DialOption, 0, 5)
-	dialOpts = append(dialOpts, grpc.WithInsecure(), grpc.WithBlock(), grpc.FailOnNonTempDialError(true))
+	dialOpts = append(dialOpts,
+		grpc.WithTransportCredentials(insecure.NewCredentials()),
+		grpc.WithBlock(),
+	)
 	cfg.DialOpts = dialOpts
 	cfg.ServerOpts = serverOpts
 	return cfg
